% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sRSA_iterative_indep_pr.R
\name{RSAModelKLDiv3params_simpleRSA4TrialsIterative_pr}
\alias{RSAModelKLDiv3params_simpleRSA4TrialsIterative_pr}
\title{Simple RSA model Kullback-Leibler divergence determination
(iterative setting, independent of trial order)
\code{posterior = (1 - prior rate) $\cdot$ evidence + (prior rate) $\cdot$ prior}}
\usage{
RSAModelKLDiv3params_simpleRSA4TrialsIterative_pr(data, par1, par2, par3)
}
\arguments{
\item{data}{A matrix with data rows.

column structure: \code{[1:OC1,OC2,OC3,4:numUttOptions,7-X:TurkerSliderValues]}

\strong{1:OC1} Object 1. A value between 1 and 27.

\strong{2:OC2} Object 2. A value between 1 and 27.

\strong{3:OC3} Object 3. A value between 1 and 27.

\strong{4:numUttOptions} The number of valid utterances in the scene.

\strong{7-X:TurkerSliderValues} These columns contain the participants' slider values.}

\item{par1}{\describe{
\item{softness parameter}{The strength of "preferring one entity over others". (The larger the value the higher the tendency towards uniform liking)}
}}

\item{par2}{\describe{
\item{non-obedience parameter}{The extent to which the instruction of the speaker is obeyed by the listener.
(0 = full obedience, infinity = full instruction ignorance).}
}}

\item{par3}{\describe{
\item{prior rate parameter}{This parameter specifies how much the prior information is weighed into the speaker's decision.
}}}
}
\value{
Minimized Kullback-Leibler divergence and the optimal parameter values.

Minimized Kullback-Leibler divergence and the optimal parameters.
}
\description{
Simple RSA (iterative, independent of trial order (prior rate))

The function calculates the optimal parameter values of the free parameters by
estimating the log-likelihood of the RSA model given model parameters and data.
It also determines the actual RSA model Kullback-Leibler divergence.

2 parameter optimization considering only the available feature values present in the scene, i.e. feature values of shape, texture and color.
This function is used in the iterative dependent on the trial scenario.
}
\details{
This function is used in \code{\link{LL1_1_Iterative_pr_notObey0_pr0.5}},

\code{\link{LL1_2_Iterative_pr_pref0_pr0.5}},

\code{\link{LL1_3_Iterative_pr_pref0_notObey0}},

\code{\link{LL2_13_Iterative_pr_notObey0}},

\code{\link{LL2_13_Iterative_pr_notObey0.1}},

\code{\link{LL2_23_Iterative_pr_pref0}}.
}
